refactor(holon-viz): seed() constructor + HasVisualization z_layer/semantic_type enrichment
- TypeRelationshipGraph::seed() moves hardcoded 47-node/52-edge graph out of
  Tauri command layer into holon-viz where it belongs
- TypeNode and CytoscapeNodeData gain z_layer/semantic_type optional fields;
  21 HasVisualization impl types annotated with exact ZLayer/SemanticType values
  from ledger-core/src/iso_objects.rs
- commands.rs get_type_graph() reduced to one line; unused imports removed
- main.js gains z_layer-keyed Cytoscape CSS selectors for consistent ZLayer palette
- ledger_ops.rs append_row call site updated to TransactionRow::new() struct form
