+++ crates/holon-viz/src/cytoscape.rs
@@ -17,6 +17,12 @@ pub struct CytoscapeNodeData {
     /// Optional parent for compound graphs (Cytoscape compound nodes).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub parent: Option<String>,
+    /// `ZLayer` variant from `HasVisualization::viz_spec()`, if available.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub z_layer: Option<String>,
+    /// `SemanticType` variant from `HasVisualization::viz_spec()`, if available.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub semantic_type: Option<String>,
 }
 
 /// A single Cytoscape.js node element.
@@ -64,6 +70,8 @@ impl CytoscapeGraph {
                     label: h.label.clone(),
                     kind: format!("{:?}", h.kind),
                     parent: h.parent_id.clone(),
+                    z_layer: None,
+                    semantic_type: None,
                 },
             })
             .collect();
