+++ crates/holon-viz/src/type_graph.rs
@@ -20,6 +20,12 @@ pub struct TypeNode {
     /// Optional Cytoscape compound parent ID, typically a module or namespace node.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub parent_id: Option<String>,
+    /// `ZLayer` variant from `HasVisualization::viz_spec()`, if the type implements that trait.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub z_layer: Option<String>,
+    /// `SemanticType` variant from `HasVisualization::viz_spec()`, if the type implements that trait.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub semantic_type: Option<String>,
 }
 
 /// Supported relationship kinds between Rust types.
@@ -113,6 +119,124 @@ impl TypeRelationshipGraph {
     pub fn to_cytoscape(&self) -> CytoscapeGraph {
         CytoscapeGraph::from(self)
     }
+
+    /// Canonical seed graph derived from `HasVisualization` impls in `ledger-core`.
+    ///
+    /// Nodes annotated with `z_layer`/`semantic_type` correspond to the 21 types that
+    /// implement `HasVisualization` in `crates/ledger-core/src/iso_objects.rs`.
+    pub fn seed() -> Self {
+        let nodes = vec![
+            type_node("iso::HasVisualization", "HasVisualization", "abstract_trait"),
+            type_node("iso::VisualizationSpec", "VisualizationSpec", "contract_type"),
+            type_node("iso::ZLayer", "ZLayer", "metamodel_enum"),
+            type_node("iso::SemanticType", "SemanticType", "metamodel_enum"),
+            type_node("iso::RhaiDsl", "RhaiDsl", "dsl_contract"),
+            type_node("zlayer::Document", "Document", "z_document"),
+            type_node("zlayer::Pipeline", "Pipeline", "z_pipeline"),
+            type_node("zlayer::Constraint", "Constraint", "z_constraint"),
+            type_node("zlayer::Legal", "Legal", "z_legal"),
+            type_node("zlayer::FormalProof", "FormalProof", "z_proof"),
+            type_node("zlayer::Attestation", "Attestation", "z_attestation"),
+            typed_node("pipeline::PipelineState<Ingested>", "PipelineState<Ingested>", "pipeline_state", "Pipeline", "Pipeline"),
+            typed_node("pipeline::PipelineState<Validated>", "PipelineState<Validated>", "pipeline_state", "Pipeline", "Pipeline"),
+            typed_node("pipeline::PipelineState<Classified>", "PipelineState<Classified>", "pipeline_state", "Pipeline", "Pipeline"),
+            typed_node("pipeline::PipelineState<Reconciled>", "PipelineState<Reconciled>", "pipeline_state", "Pipeline", "Pipeline"),
+            typed_node("pipeline::PipelineState<Committed>", "PipelineState<Committed>", "pipeline_state", "Pipeline", "Pipeline"),
+            typed_node("pipeline::PipelineState<NeedsReview>", "PipelineState<NeedsReview>", "review_state", "Pipeline", "Pipeline"),
+            typed_node("validation::CommitGate", "CommitGate", "gate_type", "Pipeline", "Gate"),
+            typed_node("validation::StageResult<T>", "StageResult<T>", "validation_type", "Pipeline", "Result"),
+            typed_node("validation::Issue", "Issue", "issue_type", "Constraint", "Issue"),
+            typed_node("validation::MetaFlag", "MetaFlag", "flag_type", "Pipeline", "Flag"),
+            typed_node("constraints::VendorConstraintSet", "VendorConstraintSet", "constraint_type", "Constraint", "Constraint"),
+            typed_node("constraints::ConstraintEvaluation", "ConstraintEvaluation", "result_type", "Constraint", "Result"),
+            typed_node("constraints::InvoiceConstraintSolver", "InvoiceConstraintSolver", "solver_type", "Constraint", "Solver"),
+            typed_node("constraints::InvoiceVerification", "InvoiceVerification", "result_type", "Constraint", "Result"),
+            typed_node("legal::Jurisdiction", "Jurisdiction", "legal_type", "Legal", "Legal"),
+            typed_node("legal::LegalRule", "LegalRule", "legal_type", "Legal", "Legal"),
+            typed_node("legal::TransactionFacts", "TransactionFacts", "fact_type", "Legal", "Legal"),
+            typed_node("legal::LegalSolver", "LegalSolver", "solver_type", "Legal", "Solver"),
+            typed_node("legal::Z3Result", "Z3Result", "proof_result", "Legal", "Result"),
+            typed_node("pipeline::KasuariSolver", "KasuariSolver", "solver_type", "FormalProof", "Proof"),
+            typed_node("attest::AttestationSpec", "AttestationSpec", "attestation_type", "Attestation", "Attestation"),
+            type_node("ontology::ArtifactKind", "ArtifactKind", "ontology_enum"),
+            type_node("ontology::RelationKind", "RelationKind", "ontology_enum"),
+            type_node("ontology::OntologySnapshot", "OntologySnapshot", "ontology_snapshot"),
+            type_node("arc_kit_au::EvidenceGraph", "EvidenceGraph", "evidence_graph"),
+            type_node("arc_kit_au::NodeType", "NodeType", "ontology_enum"),
+            type_node("arc_kit_au::SourceDoc", "SourceDoc", "evidence_node"),
+            type_node("arc_kit_au::ExtractedRow", "ExtractedRow", "evidence_node"),
+            type_node("arc_kit_au::Transaction", "Transaction", "evidence_node"),
+            type_node("arc_kit_au::Classification", "Classification", "evidence_node"),
+            type_node("arc_kit_au::ModelProposal", "ModelProposal", "evidence_node"),
+            type_node("arc_kit_au::OperatorApproval", "OperatorApproval", "evidence_node"),
+            type_node("arc_kit_au::WorkbookRow", "WorkbookRow", "evidence_node"),
+            type_node("workbook::TxProjectionRow", "TxProjectionRow", "workbook_projection"),
+            type_node("classify::TaxCategory", "TaxCategory", "taxonomy_type"),
+            type_node("workflow::WorkflowToml", "WorkflowToml", "workflow_type"),
+        ];
+
+        let relationships = vec![
+            rel("iso::VisualizationSpec", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("iso::VisualizationSpec", "iso::ZLayer", TypeRelationshipKind::Contains),
+            rel("iso::VisualizationSpec", "iso::SemanticType", TypeRelationshipKind::Contains),
+            rel("iso::VisualizationSpec", "iso::RhaiDsl", TypeRelationshipKind::Contains),
+            rel("pipeline::PipelineState<Ingested>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<Validated>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<Classified>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<Reconciled>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<Committed>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<NeedsReview>", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("constraints::VendorConstraintSet", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("constraints::ConstraintEvaluation", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("constraints::InvoiceConstraintSolver", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("constraints::InvoiceVerification", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("legal::LegalRule", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("legal::LegalSolver", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("legal::Z3Result", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("validation::CommitGate", "iso::HasVisualization", TypeRelationshipKind::Implements),
+            rel("pipeline::PipelineState<Ingested>", "pipeline::PipelineState<Validated>", TypeRelationshipKind::AdvancesTo),
+            rel("pipeline::PipelineState<Validated>", "pipeline::PipelineState<Classified>", TypeRelationshipKind::AdvancesTo),
+            rel("pipeline::PipelineState<Classified>", "pipeline::PipelineState<Reconciled>", TypeRelationshipKind::AdvancesTo),
+            rel("pipeline::PipelineState<Reconciled>", "validation::CommitGate", TypeRelationshipKind::AdvancesTo),
+            rel("validation::CommitGate", "pipeline::PipelineState<Committed>", TypeRelationshipKind::AdvancesTo),
+            rel("validation::CommitGate", "pipeline::PipelineState<NeedsReview>", TypeRelationshipKind::ValidatedBy),
+            rel("constraints::VendorConstraintSet", "constraints::ConstraintEvaluation", TypeRelationshipKind::Produces),
+            rel("constraints::InvoiceConstraintSolver", "constraints::InvoiceVerification", TypeRelationshipKind::Verifies),
+            rel("constraints::ConstraintEvaluation", "validation::Issue", TypeRelationshipKind::Produces),
+            rel("validation::Issue", "validation::StageResult<T>", TypeRelationshipKind::Contains),
+            rel("validation::StageResult<T>", "validation::CommitGate", TypeRelationshipKind::ValidatedBy),
+            rel("legal::Jurisdiction", "legal::LegalRule", TypeRelationshipKind::Contains),
+            rel("legal::TransactionFacts", "legal::LegalSolver", TypeRelationshipKind::References),
+            rel("legal::LegalRule", "legal::LegalSolver", TypeRelationshipKind::References),
+            rel("legal::LegalSolver", "legal::Z3Result", TypeRelationshipKind::Verifies),
+            rel("legal::Z3Result", "validation::Issue", TypeRelationshipKind::Produces),
+            rel("pipeline::KasuariSolver", "constraints::ConstraintEvaluation", TypeRelationshipKind::Constrains),
+            rel("legal::Z3Result", "attest::AttestationSpec", TypeRelationshipKind::Attests),
+            rel("workflow::WorkflowToml", "pipeline::PipelineState<Ingested>", TypeRelationshipKind::References),
+            rel("workflow::WorkflowToml", "pipeline::PipelineState<Committed>", TypeRelationshipKind::References),
+            rel("ontology::OntologySnapshot", "ontology::ArtifactKind", TypeRelationshipKind::Contains),
+            rel("ontology::OntologySnapshot", "ontology::RelationKind", TypeRelationshipKind::Contains),
+            rel("ontology::RelationKind", "arc_kit_au::EvidenceGraph", TypeRelationshipKind::ProjectsTo),
+            rel("arc_kit_au::EvidenceGraph", "arc_kit_au::NodeType", TypeRelationshipKind::Contains),
+            rel("arc_kit_au::SourceDoc", "arc_kit_au::ExtractedRow", TypeRelationshipKind::Produces),
+            rel("arc_kit_au::ExtractedRow", "arc_kit_au::Transaction", TypeRelationshipKind::Produces),
+            rel("arc_kit_au::Transaction", "arc_kit_au::Classification", TypeRelationshipKind::ClassifiedAs),
+            rel("arc_kit_au::Classification", "arc_kit_au::ModelProposal", TypeRelationshipKind::ValidatedBy),
+            rel("arc_kit_au::ModelProposal", "arc_kit_au::OperatorApproval", TypeRelationshipKind::ValidatedBy),
+            rel("arc_kit_au::Transaction", "arc_kit_au::WorkbookRow", TypeRelationshipKind::ProjectsTo),
+            rel("arc_kit_au::WorkbookRow", "workbook::TxProjectionRow", TypeRelationshipKind::ProjectsTo),
+            rel("workbook::TxProjectionRow", "classify::TaxCategory", TypeRelationshipKind::ClassifiedAs),
+            rel("arc_kit_au::EvidenceGraph", "workbook::TxProjectionRow", TypeRelationshipKind::RecordsIn),
+            rel("zlayer::Document", "arc_kit_au::SourceDoc", TypeRelationshipKind::Contains),
+            rel("zlayer::Pipeline", "pipeline::PipelineState<Ingested>", TypeRelationshipKind::Contains),
+            rel("zlayer::Constraint", "constraints::VendorConstraintSet", TypeRelationshipKind::Contains),
+            rel("zlayer::Legal", "legal::LegalRule", TypeRelationshipKind::Contains),
+            rel("zlayer::FormalProof", "legal::Z3Result", TypeRelationshipKind::Contains),
+            rel("zlayer::Attestation", "attest::AttestationSpec", TypeRelationshipKind::Contains),
+        ];
+
+        Self::new(nodes, relationships)
+    }
 }
 
 impl From<&TypeRelationshipGraph> for CytoscapeGraph {
@@ -136,6 +260,8 @@ impl From<&TypeRelationshipGraph> for CytoscapeGraph {
                     label: node.label,
                     kind: node.kind,
                     parent: node.parent_id,
+                    z_layer: node.z_layer,
+                    semantic_type: node.semantic_type,
                 },
             })
             .collect();
@@ -166,6 +292,32 @@ impl From<TypeRelationshipGraph> for CytoscapeGraph {
     }
 }
 
+fn type_node(id: &str, label: &str, kind: &str) -> TypeNode {
+    TypeNode {
+        id: id.to_string(),
+        label: label.to_string(),
+        kind: kind.to_string(),
+        parent_id: None,
+        z_layer: None,
+        semantic_type: None,
+    }
+}
+
+fn typed_node(id: &str, label: &str, kind: &str, z_layer: &str, semantic_type: &str) -> TypeNode {
+    TypeNode {
+        id: id.to_string(),
+        label: label.to_string(),
+        kind: kind.to_string(),
+        parent_id: None,
+        z_layer: Some(z_layer.to_string()),
+        semantic_type: Some(semantic_type.to_string()),
+    }
+}
+
+fn rel(source: &str, target: &str, kind: TypeRelationshipKind) -> TypeRelationship {
+    TypeRelationship::new(source, target, kind)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -176,6 +328,8 @@ mod tests {
             label: label.to_string(),
             kind: kind.to_string(),
             parent_id: None,
+            z_layer: None,
+            semantic_type: None,
         }
     }
