+++ crates/ledgerr-host/src/bin/tauri/commands.rs
@@ -16,9 +16,7 @@ use ledgerr_host::internal_openai::{
 use ledgerr_host::settings::ChatSettings;
 
 use super::state::AppState;
-use holon_viz::{
-    CytoscapeGraph, TypeNode, TypeRelationship, TypeRelationshipGraph, TypeRelationshipKind,
-};
+use holon_viz::{CytoscapeGraph, TypeRelationshipGraph};
 
 // ── Test harness config ───────────────────────────────────────────────────────
 
@@ -649,134 +647,10 @@ pub fn get_holon_viz_graph() -> Result<CytoscapeGraph, String> {
 
 /// Return the Rust type relationship graph for the Viz panel.
 ///
-/// This renders the internal ontology/invariant mesh from the mdBook and Rust
-/// domain model: concrete type-state objects, legal/constraint solvers,
-/// evidence provenance nodes, and workbook projection surfaces.
+/// Delegates to [`TypeRelationshipGraph::seed()`] in `holon-viz`.
 #[tauri::command]
 #[specta::specta]
 pub fn get_type_graph() -> Result<CytoscapeGraph, String> {
-    let nodes = vec![
-        type_node("iso::HasVisualization", "HasVisualization", "abstract_trait"),
-        type_node("iso::VisualizationSpec", "VisualizationSpec", "contract_type"),
-        type_node("iso::ZLayer", "ZLayer", "metamodel_enum"),
-        type_node("iso::SemanticType", "SemanticType", "metamodel_enum"),
-        type_node("iso::RhaiDsl", "RhaiDsl", "dsl_contract"),
-        type_node("zlayer::Document", "Document", "z_document"),
-        type_node("zlayer::Pipeline", "Pipeline", "z_pipeline"),
-        type_node("zlayer::Constraint", "Constraint", "z_constraint"),
-        type_node("zlayer::Legal", "Legal", "z_legal"),
-        type_node("zlayer::FormalProof", "FormalProof", "z_proof"),
-        type_node("zlayer::Attestation", "Attestation", "z_attestation"),
-        type_node("pipeline::PipelineState<Ingested>", "PipelineState<Ingested>", "pipeline_state"),
-        type_node("pipeline::PipelineState<Validated>", "PipelineState<Validated>", "pipeline_state"),
-        type_node("pipeline::PipelineState<Classified>", "PipelineState<Classified>", "pipeline_state"),
-        type_node("pipeline::PipelineState<Reconciled>", "PipelineState<Reconciled>", "pipeline_state"),
-        type_node("pipeline::PipelineState<Committed>", "PipelineState<Committed>", "pipeline_state"),
-        type_node("pipeline::PipelineState<NeedsReview>", "PipelineState<NeedsReview>", "review_state"),
-        type_node("validation::CommitGate", "CommitGate", "gate_type"),
-        type_node("validation::StageResult<T>", "StageResult<T>", "validation_type"),
-        type_node("validation::Issue", "Issue", "issue_type"),
-        type_node("validation::MetaFlag", "MetaFlag", "flag_type"),
-        type_node("constraints::VendorConstraintSet", "VendorConstraintSet", "constraint_type"),
-        type_node("constraints::ConstraintEvaluation", "ConstraintEvaluation", "result_type"),
-        type_node("constraints::InvoiceConstraintSolver", "InvoiceConstraintSolver", "solver_type"),
-        type_node("constraints::InvoiceVerification", "InvoiceVerification", "result_type"),
-        type_node("legal::Jurisdiction", "Jurisdiction", "legal_type"),
-        type_node("legal::LegalRule", "LegalRule", "legal_type"),
-        type_node("legal::TransactionFacts", "TransactionFacts", "fact_type"),
-        type_node("legal::LegalSolver", "LegalSolver", "solver_type"),
-        type_node("legal::Z3Result", "Z3Result", "proof_result"),
-        type_node("pipeline::KasuariSolver", "KasuariSolver", "solver_type"),
-        type_node("attest::AttestationSpec", "AttestationSpec", "attestation_type"),
-        type_node("ontology::ArtifactKind", "ArtifactKind", "ontology_enum"),
-        type_node("ontology::RelationKind", "RelationKind", "ontology_enum"),
-        type_node("ontology::OntologySnapshot", "OntologySnapshot", "ontology_snapshot"),
-        type_node("arc_kit_au::EvidenceGraph", "EvidenceGraph", "evidence_graph"),
-        type_node("arc_kit_au::NodeType", "NodeType", "ontology_enum"),
-        type_node("arc_kit_au::SourceDoc", "SourceDoc", "evidence_node"),
-        type_node("arc_kit_au::ExtractedRow", "ExtractedRow", "evidence_node"),
-        type_node("arc_kit_au::Transaction", "Transaction", "evidence_node"),
-        type_node("arc_kit_au::Classification", "Classification", "evidence_node"),
-        type_node("arc_kit_au::ModelProposal", "ModelProposal", "evidence_node"),
-        type_node("arc_kit_au::OperatorApproval", "OperatorApproval", "evidence_node"),
-        type_node("arc_kit_au::WorkbookRow", "WorkbookRow", "evidence_node"),
-        type_node("workbook::TxProjectionRow", "TxProjectionRow", "workbook_projection"),
-        type_node("classify::TaxCategory", "TaxCategory", "taxonomy_type"),
-        type_node("workflow::WorkflowToml", "WorkflowToml", "workflow_type"),
-    ];
-
-    let relationships = vec![
-        rel("iso::VisualizationSpec", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("iso::VisualizationSpec", "iso::ZLayer", TypeRelationshipKind::Contains),
-        rel("iso::VisualizationSpec", "iso::SemanticType", TypeRelationshipKind::Contains),
-        rel("iso::VisualizationSpec", "iso::RhaiDsl", TypeRelationshipKind::Contains),
-        rel("pipeline::PipelineState<Ingested>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<Validated>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<Classified>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<Reconciled>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<Committed>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<NeedsReview>", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("constraints::VendorConstraintSet", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("constraints::ConstraintEvaluation", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("constraints::InvoiceConstraintSolver", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("constraints::InvoiceVerification", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("legal::LegalRule", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("legal::LegalSolver", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("legal::Z3Result", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("validation::CommitGate", "iso::HasVisualization", TypeRelationshipKind::Implements),
-        rel("pipeline::PipelineState<Ingested>", "pipeline::PipelineState<Validated>", TypeRelationshipKind::AdvancesTo),
-        rel("pipeline::PipelineState<Validated>", "pipeline::PipelineState<Classified>", TypeRelationshipKind::AdvancesTo),
-        rel("pipeline::PipelineState<Classified>", "pipeline::PipelineState<Reconciled>", TypeRelationshipKind::AdvancesTo),
-        rel("pipeline::PipelineState<Reconciled>", "validation::CommitGate", TypeRelationshipKind::AdvancesTo),
-        rel("validation::CommitGate", "pipeline::PipelineState<Committed>", TypeRelationshipKind::AdvancesTo),
-        rel("validation::CommitGate", "pipeline::PipelineState<NeedsReview>", TypeRelationshipKind::ValidatedBy),
-        rel("constraints::VendorConstraintSet", "constraints::ConstraintEvaluation", TypeRelationshipKind::Produces),
-        rel("constraints::InvoiceConstraintSolver", "constraints::InvoiceVerification", TypeRelationshipKind::Verifies),
-        rel("constraints::ConstraintEvaluation", "validation::Issue", TypeRelationshipKind::Produces),
-        rel("validation::Issue", "validation::StageResult<T>", TypeRelationshipKind::Contains),
-        rel("validation::StageResult<T>", "validation::CommitGate", TypeRelationshipKind::ValidatedBy),
-        rel("legal::Jurisdiction", "legal::LegalRule", TypeRelationshipKind::Contains),
-        rel("legal::TransactionFacts", "legal::LegalSolver", TypeRelationshipKind::References),
-        rel("legal::LegalRule", "legal::LegalSolver", TypeRelationshipKind::References),
-        rel("legal::LegalSolver", "legal::Z3Result", TypeRelationshipKind::Verifies),
-        rel("legal::Z3Result", "validation::Issue", TypeRelationshipKind::Produces),
-        rel("pipeline::KasuariSolver", "constraints::ConstraintEvaluation", TypeRelationshipKind::Constrains),
-        rel("legal::Z3Result", "attest::AttestationSpec", TypeRelationshipKind::Attests),
-        rel("workflow::WorkflowToml", "pipeline::PipelineState<Ingested>", TypeRelationshipKind::References),
-        rel("workflow::WorkflowToml", "pipeline::PipelineState<Committed>", TypeRelationshipKind::References),
-        rel("ontology::OntologySnapshot", "ontology::ArtifactKind", TypeRelationshipKind::Contains),
-        rel("ontology::OntologySnapshot", "ontology::RelationKind", TypeRelationshipKind::Contains),
-        rel("ontology::RelationKind", "arc_kit_au::EvidenceGraph", TypeRelationshipKind::ProjectsTo),
-        rel("arc_kit_au::EvidenceGraph", "arc_kit_au::NodeType", TypeRelationshipKind::Contains),
-        rel("arc_kit_au::SourceDoc", "arc_kit_au::ExtractedRow", TypeRelationshipKind::Produces),
-        rel("arc_kit_au::ExtractedRow", "arc_kit_au::Transaction", TypeRelationshipKind::Produces),
-        rel("arc_kit_au::Transaction", "arc_kit_au::Classification", TypeRelationshipKind::ClassifiedAs),
-        rel("arc_kit_au::Classification", "arc_kit_au::ModelProposal", TypeRelationshipKind::ValidatedBy),
-        rel("arc_kit_au::ModelProposal", "arc_kit_au::OperatorApproval", TypeRelationshipKind::ValidatedBy),
-        rel("arc_kit_au::Transaction", "arc_kit_au::WorkbookRow", TypeRelationshipKind::ProjectsTo),
-        rel("arc_kit_au::WorkbookRow", "workbook::TxProjectionRow", TypeRelationshipKind::ProjectsTo),
-        rel("workbook::TxProjectionRow", "classify::TaxCategory", TypeRelationshipKind::ClassifiedAs),
-        rel("arc_kit_au::EvidenceGraph", "workbook::TxProjectionRow", TypeRelationshipKind::RecordsIn),
-        rel("zlayer::Document", "arc_kit_au::SourceDoc", TypeRelationshipKind::Contains),
-        rel("zlayer::Pipeline", "pipeline::PipelineState<Ingested>", TypeRelationshipKind::Contains),
-        rel("zlayer::Constraint", "constraints::VendorConstraintSet", TypeRelationshipKind::Contains),
-        rel("zlayer::Legal", "legal::LegalRule", TypeRelationshipKind::Contains),
-        rel("zlayer::FormalProof", "legal::Z3Result", TypeRelationshipKind::Contains),
-        rel("zlayer::Attestation", "attest::AttestationSpec", TypeRelationshipKind::Contains),
-    ];
-
-    Ok(TypeRelationshipGraph::new(nodes, relationships).to_cytoscape())
+    Ok(TypeRelationshipGraph::seed().to_cytoscape())
 }
-
-fn rel(source: &str, target: &str, kind: TypeRelationshipKind) -> TypeRelationship {
-    TypeRelationship::new(source, target, kind)
-}
-
-fn type_node(id: &str, label: &str, kind: &str) -> TypeNode {
-    TypeNode {
-        id: id.to_string(),
-        label: label.to_string(),
-        kind: kind.to_string(),
-        parent_id: None,
-    }
 }
